+++ model_chat_channel.go
@@ -1,6 +1,6 @@
 package bitmexgo
 
 type ChatChannel struct {
-	Id   float32 `json:"id,omitempty"`
-	Name string  `json:"name"`
+	Id   int    `json:"id,omitempty"`
+	Name string `json:"name"`
 }
