+++ model_connected_users.go
@@ -1,6 +1,6 @@
 package bitmexgo
 
 type ConnectedUsers struct {
-	Users float32 `json:"users,omitempty"`
-	Bots  float32 `json:"bots,omitempty"`
+	Users int `json:"users,omitempty"`
+	Bots  int `json:"bots,omitempty"`
 }
