+++ model_liquidation.go
@@ -6,5 +6,5 @@ type Liquidation struct {
 	Symbol    string  `json:"symbol,omitempty"`
 	Side      string  `json:"side,omitempty"`
 	Price     float64 `json:"price,omitempty"`
-	LeavesQty float32 `json:"leavesQty,omitempty"`
+	LeavesQty int     `json:"leavesQty,omitempty"`
 }
