+++ model_order.go
@@ -9,13 +9,13 @@ type Order struct {
 	OrderID               string    `json:"orderID"`
 	ClOrdID               string    `json:"clOrdID,omitempty"`
 	ClOrdLinkID           string    `json:"clOrdLinkID,omitempty"`
-	Account               float32   `json:"account,omitempty"`
+	Account               int       `json:"account,omitempty"`
 	Symbol                string    `json:"symbol,omitempty"`
 	Side                  string    `json:"side,omitempty"`
 	SimpleOrderQty        float64   `json:"simpleOrderQty,omitempty"`
-	OrderQty              float32   `json:"orderQty,omitempty"`
+	OrderQty              int       `json:"orderQty,omitempty"`
 	Price                 float64   `json:"price,omitempty"`
-	DisplayQty            float32   `json:"displayQty,omitempty"`
+	DisplayQty            int       `json:"displayQty,omitempty"`
 	StopPx                float64   `json:"stopPx,omitempty"`
 	PegOffsetValue        float64   `json:"pegOffsetValue,omitempty"`
 	PegPriceType          string    `json:"pegPriceType,omitempty"`
@@ -31,9 +31,9 @@ type Order struct {
 	WorkingIndicator      bool      `json:"workingIndicator,omitempty"`
 	OrdRejReason          string    `json:"ordRejReason,omitempty"`
 	SimpleLeavesQty       float64   `json:"simpleLeavesQty,omitempty"`
-	LeavesQty             float32   `json:"leavesQty,omitempty"`
+	LeavesQty             int       `json:"leavesQty,omitempty"`
 	SimpleCumQty          float64   `json:"simpleCumQty,omitempty"`
-	CumQty                float32   `json:"cumQty,omitempty"`
+	CumQty                int       `json:"cumQty,omitempty"`
 	AvgPx                 float64   `json:"avgPx,omitempty"`
 	MultiLegReportingType string    `json:"multiLegReportingType,omitempty"`
 	Text                  string    `json:"text,omitempty"`
