+++ model_order_book_l2.go
@@ -2,8 +2,8 @@ package bitmexgo
 
 type OrderBookL2 struct {
 	Symbol string  `json:"symbol"`
-	Id     float32 `json:"id"`
+	Id     int     `json:"id"`
 	Side   string  `json:"side"`
-	Size   float32 `json:"size,omitempty"`
+	Size   int     `json:"size,omitempty"`
 	Price  float64 `json:"price,omitempty"`
 }
