+++ model_settlement.go
@@ -12,7 +12,7 @@ type Settlement struct {
 	SettledPrice          float64   `json:"settledPrice,omitempty"`
 	OptionStrikePrice     float64   `json:"optionStrikePrice,omitempty"`
 	OptionUnderlyingPrice float64   `json:"optionUnderlyingPrice,omitempty"`
-	Bankrupt              float32   `json:"bankrupt,omitempty"`
-	TaxBase               float32   `json:"taxBase,omitempty"`
+	Bankrupt              int       `json:"bankrupt,omitempty"`
+	TaxBase               int       `json:"taxBase,omitempty"`
 	TaxRate               float64   `json:"taxRate,omitempty"`
 }
