+++ model_stats_history.go
@@ -8,6 +8,6 @@ type StatsHistory struct {
 	Date       time.Time `json:"date"`
 	RootSymbol string    `json:"rootSymbol"`
 	Currency   string    `json:"currency,omitempty"`
-	Volume     float32   `json:"volume,omitempty"`
-	Turnover   float32   `json:"turnover,omitempty"`
+	Volume     int       `json:"volume,omitempty"`
+	Turnover   int       `json:"turnover,omitempty"`
 }
